@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "diagrams"
-version = "0.23.1"
+version = "0.24.1"
 description = "Diagram as Code"
 license = "MIT"
 authors = ["mingrammer <mingrammer@gmail.com>"]
@@ -9,19 +9,22 @@ homepage = "https://diagrams.mingrammer.com"
 repository = "https://github.com/mingrammer/diagrams"
 include = ["resources/**/*"]
 
+[tool.poetry.scripts]
+diagrams="diagrams.cli:main"
+
 [tool.poetry.dependencies]
-python = "^3.7"
-graphviz = ">=0.13.2,<0.20.0"
+python = "^3.9"
+graphviz = ">=0.13.2,<0.21.0"
 jinja2 = ">=2.10,<4.0"
-contextvars = { version = "^2.4", python = "~3.6" }
-typed-ast = "^1.5.4"
+typed-ast = {version="^1.5.5", markers="python_version<'3.8'"}
+pre-commit = "^4.0.1"
 
 [tool.poetry.dev-dependencies]
-pytest = "^7.0"
-pylint = "^2.7"
-rope = "^0.14.0"
-isort = "^4.3"
-black = "^22.10.0"
+pytest = "^8.3"
+pylint = "^3.3"
+rope = "^1.13"
+isort = "^5.13"
+black = "^24.4"
 
 [tool.black]
-line-length = 120
+line-length=120
