@@ -1,5 +1,5 @@
 """
-OnPrem provides a set of general on-premise services.
+OnPrem provides a set of general on-premises services.
 """
 
 from diagrams import Node
@@ -10,3 +10,7 @@ class _OnPrem(Node):
     _icon_dir = "resources/onprem"
 
     fontcolor = "#ffffff"
+
+
+class OnPrem(_OnPrem):
+    _icon = "onprem.png"
