@@ -11,6 +11,7 @@ import subprocess
 import sys
 
 import config as cfg
+
 from . import resource_dir
 
 _usage = "Usage: resource.py <cmd> <pvd>"
@@ -30,7 +31,7 @@ def cleaner_aws(f):
     f = f.replace("-light-bg", "")
     for p in cfg.FILE_PREFIXES["aws"]:
         if f.startswith(p):
-            f = f[len(p) :]
+            f = f[len(p):]
             break
     return f.lower()
 
@@ -41,7 +42,7 @@ def cleaner_azure(f):
     f = "-".join(f.split())
     for p in cfg.FILE_PREFIXES["azure"]:
         if f.startswith(p):
-            f = f[len(p) :]
+            f = f[len(p):]
             break
     return f.lower()
 
@@ -51,7 +52,7 @@ def cleaner_gcp(f):
     f = "-".join(f.split())
     for p in cfg.FILE_PREFIXES["gcp"]:
         if f.startswith(p):
-            f = f[len(p) :]
+            f = f[len(p):]
             break
     return f.lower()
 
@@ -61,7 +62,7 @@ def cleaner_ibm(f):
     f = "-".join(f.split())
     for p in cfg.FILE_PREFIXES["ibm"]:
         if f.startswith(p):
-            f = f[len(p) :]
+            f = f[len(p):]
             break
     return f.lower()
 
@@ -71,7 +72,7 @@ def cleaner_firebase(f):
     f = "-".join(f.split())
     for p in cfg.FILE_PREFIXES["firebase"]:
         if f.startswith(p):
-            f = f[len(p) :]
+            f = f[len(p):]
             break
     return f.lower()
 
@@ -80,15 +81,16 @@ def cleaner_k8s(f):
     f = f.replace("-256", "")
     for p in cfg.FILE_PREFIXES["k8s"]:
         if f.startswith(p):
-            f = f[len(p) :]
+            f = f[len(p):]
             break
     return f.lower()
+
 
 def cleaner_digitalocean(f):
     f = f.replace("-32", "")
     for p in cfg.FILE_PREFIXES["digitalocean"]:
         if f.startswith(p):
-            f = f[len(p) :]
+            f = f[len(p):]
             break
     return f.lower()
 
@@ -96,7 +98,7 @@ def cleaner_digitalocean(f):
 def cleaner_alibabacloud(f):
     for p in cfg.FILE_PREFIXES["alibabacloud"]:
         if f.startswith(p):
-            f = f[len(p) :]
+            f = f[len(p):]
             break
     return f.lower()
 
@@ -106,7 +108,7 @@ def cleaner_oci(f):
     f = f.replace("_", "-")
     for p in cfg.FILE_PREFIXES["oci"]:
         if f.startswith(p):
-            f = f[len(p) :]
+            f = f[len(p):]
             break
     return f.lower()
 
@@ -201,7 +203,8 @@ def svg2png2(pvd: str) -> None:
     def _convert(base: str, path: str):
         path_src = os.path.join(base, path)
         path_dest = path_src.replace(".svg", ".png")
-        subprocess.run([cfg.CMD_SVG2PNG_IM, *cfg.CMD_SVG2PNG_IM_OPTS, path_src, path_dest])
+        subprocess.run([cfg.CMD_SVG2PNG_IM, *
+                        cfg.CMD_SVG2PNG_IM_OPTS, path_src, path_dest])
         subprocess.run(["rm", path_src])
 
     for root, _, files in os.walk(resource_dir(pvd)):
