@@ -3,14 +3,14 @@ import random
 import string
 import unittest
 
-from diagrams import Diagram
-from diagrams import setcluster, setdiagram
-from diagrams.c4 import Person, Container, Database, System, SystemBoundary, Relationship
+from diagrams import Diagram, setcluster, setdiagram
+from diagrams.c4 import Container, Database, Person, Relationship, System, SystemBoundary
 
 
 class C4Test(unittest.TestCase):
     def setUp(self):
-        self.name = "diagram-" + "".join([random.choice(string.hexdigits) for n in range(7)])
+        self.name = "diagram-" + \
+            "".join([random.choice(string.hexdigits) for n in range(7)]).lower()
 
     def tearDown(self):
         setdiagram(None)
@@ -23,8 +23,14 @@ class C4Test(unittest.TestCase):
     def test_nodes(self):
         with Diagram(name=self.name, show=False):
             person = Person("person", "A person.")
-            container = Container("container", "Java application", "The application.")
-            database = Database("database", "Oracle database", "Stores information.")
+            container = Container(
+                "container",
+                "Java application",
+                "The application.")
+            database = Database(
+                "database",
+                "Oracle database",
+                "Stores information.")
 
     def test_external_nodes(self):
         with Diagram(name=self.name, show=False):
